+++ src/common.cpp
@@ -93,7 +93,30 @@ void run_original(const float data[], const int len, float& sum_value, float& ma
 
 void run_speedup(const float data[], const int len, float& sum_value, float& max_value, float result[]) {
     // TODO: Distinguish client and server mode.
+#if TIME_RECORD
+    timespec start, end;
+    double time_consumed;
+
+    clock_gettime(CLOCK_REALTIME, &start);
+    sum_value = sumSpeedUp(data, len);
+    clock_gettime(CLOCK_REALTIME, &end);
+    time_consumed = end.tv_sec - start.tv_sec + (end.tv_nsec - start.tv_nsec) / 1e9;
+    std::cout << "sum time consumed: " << time_consumed << "s" << std::endl;
+
+    clock_gettime(CLOCK_REALTIME, &start);
+    max_value = maxSpeedUp(data, len);
+    clock_gettime(CLOCK_REALTIME, &end);
+    time_consumed = end.tv_sec - start.tv_sec + (end.tv_nsec - start.tv_nsec) / 1e9;
+    std::cout << "max time consumed: " << time_consumed << "s" << std::endl;
+
+    clock_gettime(CLOCK_REALTIME, &start);
+    sortSpeedUp(data, len, result);
+    clock_gettime(CLOCK_REALTIME, &end);
+    time_consumed = end.tv_sec - start.tv_sec + (end.tv_nsec - start.tv_nsec) / 1e9;
+    std::cout << "sort time consumed: " << time_consumed << "s" << std::endl;
+#else
     sum_value = sumSpeedUp(data, len);
     max_value = maxSpeedUp(data, len);
-    // sortSpeedUp(data, len, result);
+    // sort(data, len, result);
+#endif
 }